Functions shuffled into moodDetection.py
