+++ moodDetection.py
@@ -1,6 +1,8 @@
 from deepface import DeepFace
 import cv2
 
+global available_emotions
+available_emotions = set(["Happy", "Sad", "Angry", "Fear", "Surprise", "Neutral"])
 
 def detectEmotion(image: cv2.imread) -> str:
     """
@@ -14,3 +16,34 @@ def detectEmotion(image: cv2.imread) -> str:
     """
     emotion = DeepFace.analyze(image, actions=["emotion"])
     return emotion[0]["dominant_emotion"]
+
+
+def combine_emotions(t_emotions: dict, i_emotions: dict) -> dict:
+    """
+    Combines the emotions from the text and the image
+    :param t_emotions: the emotions from the text
+    :param i_emotions: the emotions from the image
+    :return: the combined emotions
+    """
+
+    norm_emotions = dict()
+    for t_emotion, t_value in t_emotions.items():
+        norm_emotions[t_emotion] = (t_value + i_emotions[t_emotion]) / 2
+
+    return norm_emotions
+
+
+def interpret_emotion(emotion: dict) -> str:
+    """
+    Interprets the emotion
+    :param emotion: the emotion to interpret
+    :return: the emotion that is the strongest
+    """
+    if "Neutral" not in emotion:
+        emotion["Neutral"] = 0.0
+    max_emotion = ["Neutral", 0.0]
+    for emo in available_emotions:
+        if emotion[emo] > max_emotion[1]:
+            max_emotion = emo, emotion[emo]
+
+    return max_emotion[0] if max_emotion[1] > 0.2 else "Neutral"
